refactor: rename `/design` URLs to `/work` URLs
