+++ pages/work/[...project].scss
